Phase 1.3: Move deprecated functions to legacy directory
- Created star_field_deprecated.rs in legacy/wasm/src/
- Moved all deprecated raw pointer functions from main star_field.rs
- Clean main API - deprecated functions no longer exported
- Updated legacy/wasm/src/lib.rs to include new module
- Zero impact on active codebase, all functionality preserved
- Clear separation between legacy and safe, actively used code
- Build and exports confirmed working correctly
